Updated files

+++ backend/app/controllers/coupon-controller.js
@@ -1,9 +1,19 @@
 const Coupon=require('../models/coupon-model')
+const {couponValidationSchema}=require('../validations/coupon-validation')
 const couponCltr={}
 couponCltr.create=async(req,res)=>{
     const body=req.body;
+    const {error,value}=couponValidationSchema.validate(body,{abortEarly:true})
+    if(error){
+        return res.status(400).json({error:error.message})
+    }
     try{
-        const coupon=await Coupon.create(body)
+        const couponExist=await Coupon.findOne({code:value.code,organiserId:req.userId})
+        if(couponExist){
+            return res.status(400).json({err:"already exists"})
+        }
+        const coupon=new Coupon(value)
+        await coupon.save();
         res.json(coupon)
     }catch(err){
         console.log(err)
@@ -23,10 +33,17 @@ couponCltr.list=async(req,res)=>{
 couponCltr.update=async(req,res)=>{
     const id=req.params.id
     const body=req.body
-  try{
-   const coupon=await Coupon.findByIdAndUpdate(id,body,{new:true})
-    res.json(coupon)
-  }catch(err){
+    const {error,value}=couponValidationSchema.validate(body,{abortEarly:true})
+    if(error){
+        return res.status(400).json({error:error.details})
+    }
+    if (!id || !id.match(/^[0-9a-fA-F]{24}$/)) {
+      return res.status(400).json({ err: "Invalid coupon ID" })
+    }
+    try{
+        const coupon=await Coupon.findByIdAndUpdate({_id:id,organiserId:req.userId},value,{new:true})
+        res.json(coupon)
+     }catch(err){
        console.log(err)
         res.status(500).json({err:"Something went wrong"})
   }
@@ -34,8 +51,11 @@ couponCltr.update=async(req,res)=>{
 
 couponCltr.remove=async(req,res)=>{
     const id=req.params.id
+        if (!id || !id.match(/^[0-9a-fA-F]{24}$/)) {
+      return res.status(400).json({ err: "Invalid coupon ID" })
+    }
     try{
-        const coupon=await Coupon.findByIdAndDelete(id)
+        const coupon=await Coupon.findByIdAndDelete({_id:id,organiserId:req.userId})
         res.json(coupon)
     }catch(err){
         console.log(err)
@@ -43,3 +63,4 @@ couponCltr.remove=async(req,res)=>{
     }
 }
 module.exports=couponCltr
+
